Add methods for linking a keyring to another keyring
This is an important keyctl operation that can have implications on
whether a key is readable and shows up in Search operations.

It may not be immediately be apparent that this is a supported operation
when calling KEYCTL_LINK, you have to dig through the docs to find this.

man 7 keyrings:

> As previously mentioned, keyrings are a special type of key that
contains links to other keys (which may include other keyrings). Keys
may be linked to by multiple keyrings.